% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_levene_test}
\alias{helper_levene_test}
\title{Helper for Levene test}
\usage{
helper_levene_test(x, g, paired)
}
\arguments{
\item{x}{a numeric vector.}

\item{g}{a factor.}

\item{paired}{a logical value.}
}
\value{
a list indicating the test and its p-value.
}
\description{
It performs the Levene test to test homoscedasticity.
}
