% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_missing}
\alias{helper_missing}
\title{Number of missing observations}
\usage{
helper_missing(var, ...)
}
\arguments{
\item{var}{a numeric vector.}

\item{...}{additional input arguments that may be used when creating your own function.}
}
\value{
a list with the first element \code{name} as the measure name and the
second element as the \code{value} for a given variable.
}
\description{
It calculates the number of missing observations.
}
\details{
This function can be modified by the user,
but input and output should be kept the same.
}
