% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_sf_test}
\alias{helper_sf_test}
\title{Percentages and frequencies}
\usage{
helper_sf_test(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
a list with the first element \code{name} as the measure name and the
second element as the \code{value} for a given variable.
}
\description{
It calculates percentages and frequencies, concatenating them to
present on a table.
}
\details{
This function can be modified by the user,
but input and output should be kept the same.
}
