% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{nt_barplot}
\alias{nt_barplot}
\title{Barplot}
\usage{
nt_barplot(
  data,
  group = NULL,
  ylab = "Percent (\%)",
  save = FALSE,
  fig.height = 5,
  fig.width = 5,
  std_fun = std_barplot,
  std_fun_group = std_barplot_group
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{an optional data frame with the group variable.}

\item{ylab}{a character value specifying the y axis label.}

\item{save}{a logical value indicating whether the output
should be saved as a jpeg file.}

\item{fig.height}{a numeric value indicating the height (in) of the file.}

\item{fig.width}{a numeric value indicating the width (in) of the file.}

\item{std_fun}{a function to plot a barplot when \code{group = NULL}.
It must follow the same structure of \code{\link{std_barplot}}.}

\item{std_fun_group}{a function to plot a dotplot when \code{group}
is provided. It must follow the same structure of
\code{\link{std_barplot_group}}.}
}
\value{
a list of ggplot objects with each item named by the column names from
\code{var}.
}
\description{
Plot barplot for several variables.
}
\details{
The functions \code{\link{std_barplot}} and
\code{\link{std_barplot_group}} are stardard functions that can be
modified by the user in order to customize the barplots a prior.
The plots also can be modified a posterior as a regular ggplot object.
See \code{\link[ggplot2]{geom_bar}}, \code{\link{std_barplot}} and
\code{\link{std_barplot_group}}.
}
\examples{
library(dplyr)
library(magrittr)
data(iris)

vars <- iris \%>\%
transmute(Species = Species,
         Sepal.Length.C = ifelse(Sepal.Length > 5, "> 5", "<= 5"),
         Sepal.Width.C = ifelse(Sepal.Width > 3.5, "> 3.5", "<= 3.5"))
vars \%>\% nt_barplot(group = Species)

}
