% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{nt_compare_tg}
\alias{nt_compare_tg}
\title{Compare two groups}
\usage{
nt_compare_tg(
  data,
  group,
  alternative = "two.sided",
  norm.test = helper_sf_test,
  var.test = helper_levene_test,
  qt.test = list(helper_student_t, helper_welch_t, helper_mann_whitney,
    helper_brunner_munzel),
  paired = FALSE,
  conf.level = 0.95,
  format = TRUE,
  digits.ci = 3,
  digits.p = 5,
  save = FALSE,
  file = "nt_compare_tg"
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{a data frame with the group variable.}

\item{alternative}{a character value indicating the alternative hypothesis,
must be one of "two.sided", "greater" or "less".}

\item{norm.test}{a function with a numeric vector as input and a list as output containing an object named \code{p.value} similar to \link[ntimes]{helper_sf_test}.}

\item{var.test}{a function with a numeric vector, group vector and paired logical variable as input and a list as output containing an object named \code{p.value} similar to \link[ntimes]{helper_levene_test}.}

\item{qt.test}{a list of functions for four possible cases: (1) normality and homoscedasticity,
(2) normality and heteroscedasticity, (3) non-normality and homoscedasticity and (4) normality and heteroscedasticity.}

\item{paired}{a logical value indicating whether a paired test should be used.}

\item{conf.level}{a character value specifying the confidence level of the confidence interval for
the difference between the two groups.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{digits.ci}{the number of digits to present the confidence intervals.}

\item{digits.p}{the number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character value indicating the name of output file in csv format to be saved.}
}
\description{
Performing comparisons between two groups.
}
\examples{
library(dplyr)
library(forcats)
data(iris)

iris \%>\% filter(Species != "setosa") \%>\%
mutate(Species = fct_drop(Species)) \%>\%
nt_compare_tg(group = Species)

}
