% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{nt_dotplot}
\alias{nt_dotplot}
\title{Dotplot}
\usage{
nt_dotplot(
  data,
  group = NULL,
  binwidth = 1,
  save = FALSE,
  fig.height = 5,
  fig.width = 5,
  std_fun = std_dotplot,
  std_fun_group = std_dotplot_group
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{an optional data frame with the group variable.}

\item{binwidth}{a numerical vector specifying the bin width for each variable.}

\item{save}{a logical value indicating whether the output
should be saved as a jpeg file.}

\item{fig.height}{a numeric value indicating the height (in) of the file.}

\item{fig.width}{a numeric value indicating the width (in) of the file.}

\item{std_fun}{a function to plot a dotplot when \code{group = NULL}.
It must follow the same structure of the function \code{\link{std_dotplot}}.}

\item{std_fun_group}{a function to plot a dotplot when \code{group}
is provided. It must follow the same structure of the function
\code{\link{std_dotplot_group}}.}
}
\value{
a list of ggplot objects with each item named by the column names from
\code{var}.
}
\description{
Plot dotplot for several variables.
}
\details{
The functions std_dotplot and std_dotplot_group can be
modified by the user in order to customize the dotplots a prior.
The plots also can be modified a posterior as a regular ggplot object.
See \code{\link[ggplot2]{geom_dotplot}}, \code{\link{std_dotplot}} and
\code{\link{std_dotplot_group}}.
}
\examples{
data(iris)

iris |> nt_dotplot(group = Species, binwidth = 0.1)

}
