% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_tables.R
\name{nt_join_tables}
\alias{nt_join_tables}
\title{Join tables}
\usage{
nt_join_tables(
  tab.x,
  tab.y,
  digits = 3,
  alpha = 0.05,
  save = FALSE,
  file = "table"
)
}
\arguments{
\item{tab.x}{a data frame with attribute \code{ntimes}.}

\item{tab.y}{a data frame with attribute \code{ntimes}.}

\item{digits}{a numeric value specifying the number of digits for the p values.}

\item{alpha}{a numeric value specifying the significance level to indicate statistically significant multiple comparisons.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character value indicating the name of output file in csv format to be saved.}
}
\value{
A data frame containing \code{tab.x} and \code{tab.y}.
}
\description{
Join tables resulting from other functions of the package \code{ntimes}.
}
\examples{
library(dplyr)
library(magrittr)

iris_nt <- iris \%>\% filter(Species != "versicolor") \%>\%
mutate(Species = droplevels(Species))
tab01 <- nt_describe(iris_nt, group = Species)
tab02 <- nt_compare_tg(iris_nt, group = Species)
tab <- nt_join_tables(tab01, tab02)

}
