% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.r
\name{nt_missing}
\alias{nt_missing}
\title{Number of missing observations}
\usage{
nt_missing(var, digits)
}
\arguments{
\item{var}{a numeric vector.}

\item{digits}{a numeric value specifying the number of digits to present the
results. It is not used for the number of missing observations.}
}
\value{
a list with the first element \code{name} as the measure name and the
second element as the \code{value} for a given variable.
}
\description{
It calculates the number of missing observations.
}
\details{
This function can be modified by the user,
but input and output should be kept the same.
}
