% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{nt_multiple_cox}
\alias{nt_multiple_cox}
\title{Proportional Hazards Cox regression table}
\usage{
nt_multiple_cox(
  fit,
  ci_type = "wald",
  contrast_qt = "one-unit",
  user_contrast = NULL,
  user_contrast_interaction = NULL,
  table_reference = TRUE,
  format = TRUE,
  labels = NULL,
  digits = 2,
  digits_p = 3,
  save = FALSE,
  file = "nt_multiple_cox"
)
}
\arguments{
\item{fit}{a coxph object.}

\item{ci_type}{a character value indicating the procedure to calculate confidence intervals: likelihood ratio (\code{lr}) or wald (\code{wald}).}

\item{user_contrast}{a variable named list of numerical vectors indicating contrast for a covariate.}

\item{user_contrast_interaction}{a variable named list of numerical vectors indicating a contrast for interaction.}

\item{table_reference}{a logical value indicating whether the output should be presented with a line indicating the reference category.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{labels}{a list of labels with components given by their variable names.}

\item{digits}{a numerical value defining of digits to present the results.}

\item{digits_p}{a numerical value defining number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\description{
Tabulating results from fitted Proportional Hazards Cox models.
}
\examples{
library(survival)
library(dplyr)

data(ovarian)
dt <- ovarian |> mutate(resid.ds = factor(resid.ds,
                                               levels = 1:2,
                                               labels = c("no", "yes")),
                             ecog.ps = factor(ecog.ps,
                                              levels = 1:2,
                                              labels  = c("I", "II")),
                             rx = factor(rx,
                                         levels = 1:2,
                                         labels = c("t1", "t2")))


fit <- coxph(Surv(futime, fustat) ~ age + ecog.ps*rx, data = dt)

nt_multiple_cox(fit)

}
