% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{nt_norm_test}
\alias{nt_norm_test}
\title{Normality tests}
\usage{
nt_norm_test(data, group = NULL, test = "sf", digits = 3, pvalue.plot = TRUE)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{an optional character indicating the group variable.}

\item{test}{a character value specifying the normality test to be performed.
The options are Anderson-Darling (\code{ad}), Shapiro-Francia (\code{"sf"}),
Kolmogorov-Smirnov (\code{ks}),  Cramer-vonMises (\code{cvm}) and
Pearson (\code{ps}). The default is Shapiro-Francia (\code{"sf"}).}

\item{digits}{a integer value indicating the numer of decimal places.}

\item{pvalue.plot}{a logical value indicating if the p-value
should be presented in the Quantile-Quantile plot.}
}
\value{
\code{tab} a table of p-values for each of the five normality tests.

\code{plot} a Quantile-Quantile plot.
}
\description{
Perform Anderson-Darling, Shapiro-Francia,
Kolmogorov-Smirnov, Cramer-vonMises, and Pearson normality tests
for several variables. In addition, it also plots a Quantile-Quantile plot and a histogram.
}
\details{
The function is a wrapper of \link[nortest]{ad.test},
\link[nortest]{sf.test}, \link[nortest]{lillie.test},
\link[nortest]{cvm.test}, and \link[nortest]{pearson.test}.
}
\examples{
library(magrittr)
data(iris)

iris \%>\% select(-Species) \%>\% nt_norm_test()
iris \%>\% nt_norm_test(group = Species)

}
