% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileplot.R
\name{nt_profileplot}
\alias{nt_profileplot}
\title{Profile plot}
\usage{
nt_profileplot(
  data,
  time = NULL,
  group = NULL,
  labels = NULL,
  save = FALSE,
  fig.height = 5,
  fig.width = 5,
  std_fun = std_profileplot,
  std_fun_group = std_profileplot_group
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{time}{a character value indicating the time variable.}

\item{group}{a character value indicating the group variable.}

\item{labels}{a list of labels with components given by their variable names.}

\item{save}{a logical value indicating whether the output
should be saved as a jpeg file.}

\item{fig.height}{a numeric value indicating the height (in) of the file.}

\item{fig.width}{a numeric value indicating the width (in) of the file.}

\item{std_fun}{a function to plot a profile plot when \code{group = NULL}.
It must follow the same structure of \code{\link{std_profileplot}}.}

\item{std_fun_group}{a function to plot a profile plot when \code{group}
is provided. It must follow the same structure of
\code{\link{std_profileplot_group}}.}
}
\value{
a list of ggplot objects with each item named by the column names from
\code{var}.
}
\description{
Plot profile plot for several variables.
}
\details{
The functions \code{\link{std_profileplot}} and
\code{\link{std_profileplot_group}} can be modified by the user in order to
customize the profile plots a prior.
The plots also can be modified a posterior as a regular ggplot object.
See \code{\link[ggplot2]{geom_errorbar}}, \code{\link{std_profileplot}} and
\code{\link{std_profileplot_group}}.
}
