% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{nt_simple_cox}
\alias{nt_simple_cox}
\title{Simple Cox Regression}
\usage{
nt_simple_cox(
  data,
  time,
  status,
  ...,
  cluster = FALSE,
  strata = NULL,
  exponentiate = TRUE,
  type = "wald",
  contrast_qt = "one-unit",
  user_contrast = NULL,
  table_reference = TRUE,
  format = TRUE,
  labels = NULL,
  digits = 2,
  digits_p = 3,
  save = FALSE,
  file = "simple_cox"
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{time}{a numeric vector with the follow-up time.}

\item{status}{a numeric vector indicating status, 0 = censored, 1 = event at time.}

\item{...}{character values indicating confounding variables.}

\item{cluster}{a character vector containing the cluster variable.}

\item{strata}{a character vector containing the strata variable.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{labels}{a list of labels with components given by their variable names.}

\item{digits}{a numerical value defining of digits to present the results.}

\item{digits_p}{a numerical value defining number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}

\item{increment}{a named list indicating the magnitude of increments to calculate odds ratio for continuous covariates.}
}
\description{
Performing simple Cox regression.
}
\examples{
library(survival)
library(dplyr)
data(ovarian)

ovarian_nt <- ovarian |> mutate(resid.ds = factor(resid.ds,
                                                  levels = 1:2,
                                                  labels = c("no", "yes")),
                             ecog.ps = factor(ecog.ps,
                                              levels = 1:2,
                                              labels = c("I", "II")),
                             rx =factor(rx,
                                         levels = 1:2,
                                         labels = c("t1", "t2")))
ovarian_nt |> nt_simple_cox(time = futime, status = fustat,
                            labels = list(resid.ds = "Residual Disease",
                                          ecog.ps = "ECOG-PS",
                                          rx = "Treatment",
                                          age = "Age"))

}
