% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_regression.R
\name{nt_simple_glm}
\alias{nt_simple_glm}
\title{Simple Generalized Linear Models}
\usage{
nt_simple_glm(
  data,
  response,
  ...,
  family,
  robust.variance = NULL,
  increment = NULL,
  ci.type = "Wald",
  conf.level = 0.95,
  exponentiate = FALSE,
  format = TRUE,
  labels = NULL,
  digits = 2,
  digits.p = 3,
  save = FALSE,
  file = "simple_logistic"
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{response}{a character value indicating the response variable.}

\item{...}{character values indicating confounding variables.}

\item{family}{a character indicating family distribution. See more \code{\link[stats]{family}}.}

\item{robust.variance}{a function yielding a covariance matrix or a covariance matrix. See more \code{\link[lmtest]{coeftest}}.}

\item{increment}{a named list indicating the magnitude of increments to calculate odds ratio for continuous covariates.}

\item{ci.type}{a character value indicating the procedure to calculate confidence intervals: likelihood ratio (\code{profile}) or wald (\code{Wald})}

\item{conf.level}{a numerical value indicating the confidence level for parameters of interest.}

\item{exponentiate}{a logical value indicating whether coefficients should be exponentiated.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{labels}{a list of labels with components given by their variable names.}

\item{digits}{a numerical value defining of digits to present the results.}

\item{digits.p}{a numerical value defining number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\description{
Fit simple GLM.
}
\examples{
library(titanic)
library(dplyr)

data(titanic_train)
titanic_nt <- titanic_train |>
 mutate(Sex = factor(Sex,
                     levels = c("male", "female"),
                     labels = c("Male", "Female")),
        Pclass = factor(Pclass,
                        levels = 1:3,
                        labels = c("I", "II", "III")),
        Embarked = factor(Embarked,
                          levels = c("C", "Q", "S"),
                          labels = c("Cherbourg", "Queenstown", "Southampton")))

titanic_nt |> select(Survived, Sex, Age, Pclass, Embarked) |>
 nt_simple_glm(response = Survived, Age,
                    family = binomial(link = "logit"),
                    exponentiate = TRUE,
                    labels = list(Pclass = "Passenger class"))

}
