% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{nt_simple_logistic}
\alias{nt_simple_logistic}
\title{Simple Logistic Regression}
\usage{
nt_simple_logistic(
  data,
  response,
  ...,
  labels = NULL,
  increment = NULL,
  format = TRUE,
  digits = 2,
  digits.p = 3,
  save = FALSE,
  file = "simple_logistic"
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{response}{a character value indicating the response variable.}

\item{...}{character values indicating confounding variables.}

\item{labels}{a list of labels with components given by their variable names.}

\item{increment}{a variable named list indicating the magnitude of increments to calculate odds ratio for continuous covariates.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{digits}{a numerical value defining of digits to present the results.}

\item{digits.p}{a numerical value defining number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\description{
Performing simple Logistic regression.
}
\examples{
library(titanic)
library(magrittr)
library(dplyr)

data(titanic_train)
titanic_nt <- titanic_train \%>\%
 mutate(Sex = factor(Sex,
                     levels = c("male", "female"),
                     labels = c("Male", "Female")),
        Pclass = factor(Pclass,
                        levels = 1:3,
                        labels = c("I", "II", "III")),
        Embarked = factor(Embarked,
                          levels = c("C", "Q", "S"),
                          labels = c("Cherbourg", "Queenstown", "Southampton")))

titanic_nt \%>\% select(Survived, Sex, Age, Pclass, Embarked) \%>\%
 nt_simple_logistic(response = Survived, Age,
                    labels = list(Pclass = "Passanger class"))

}
