% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{nt_table_coxph}
\alias{nt_table_coxph}
\title{Proportional Hazards Cox regression table}
\usage{
nt_table_coxph(fit.list, data, save = FALSE, file = "nt_table_cox")
}
\arguments{
\item{fit.list}{a list of fitted models.}

\item{data}{a data frame containing the variables used to fit the models listed in fit.list.}
}
\description{
Tabulating results from fitted Proportional Hazards Cox models.
}
\examples{
library(survival)
library(magrittr)
library(dplyr)

data(ovarian)
ovarian <- ovarian \%>\% mutate(resid.ds = ql_var(resid.ds,
                                               from = 1:2,
                                               to = c("no", "yes"),
                                               label = "Residual Disease"),
                             ecog.ps = ql_var(ecog.ps,
                                              from = 1:2,
                                              to = c("I", "II"),
                                              label = "ECOG-PS"),
                             rx = ql_var(rx,
                                         from = 1:2,
                                         to = c("t1", "t2"),
                                         label = "Treatment"),
                             age = qt_var(age,
                                          label = "Age"))

fit.list <- list()

fit.list[[1]] <- coxph(Surv(futime, fustat) ~ ecog.ps, data = ovarian)
fit.list[[2]] <- coxph(Surv(futime, fustat) ~ resid.ds + rx, data = ovarian)
fit.list[[3]] <- coxph(Surv(futime, fustat) ~ age + ecog.ps*rx, data = ovarian)

nt_table_coxph(fit.list, data = ovarian)

}
