% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{put_together}
\alias{put_together}
\title{Put Together}
\usage{
put_together(
  descriptive.tab,
  test.tab,
  digits = 3,
  alpha = 0.05,
  save = FALSE,
  file = "table"
)
}
\arguments{
\item{descriptive.tab}{a data frame of class "descriptive".}

\item{test.tab}{a data frame of class "p_value".}

\item{digits}{a numeric value specifying the number of digits for the p values.}

\item{alpha}{a numeric value specifying the significance level to indicate statistically significant multiple comparisons.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character value indicating the name of output file in csv format to be saved.}
}
\value{
A data frame similar to the \code{descriptive.tab} with an additional column of
p values from the \code{p_value.tab}
}
\description{
Join a descriptive table and a p_value table.
}
\examples{
library(dplyr)
library(magrittr)

iris_nt <- iris \%>\% filter(Species != "versicolor")
tab01 <- nt_describe(iris_nt, group = Species)
tab02 <- nt_compare_tg(iris_nt, group = Species)
tab <- put_together(tab01, tab02)

}
