% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ql_var}
\alias{ql_var}
\title{Format qualitative variables}
\usage{
ql_var(var, from = NULL, to = NULL, order = NULL, label = NULL)
}
\arguments{
\item{var}{vector to be formatted.}

\item{from}{an optional vector of original levels that \code{var} might have taken. It also requires \code{to}.}

\item{to}{an optional vector of values to recode the levels of \code{from}. It also requires \code{to}.}

\item{order}{an optional vector of values to establish the level order using the same values of \code{from}.}

\item{label}{an optional character to be attributed as a label to \code{var}.}
}
\value{
a vector  of class "factor" recoded if \code{from} and \code{to} are provided,
with an attribute label if \code{label} is provided,
and reordered if \code{order} is provided.
}
\description{
Recode factors, change the order of levels and
add a label to the qualitative variable.
}
\examples{
library(dplyr)
library(magrittr)

iris <- iris \%>\% mutate(species = ql_var(Species,
from = c("setosa", "Versicolor", "virginica"),
to = c("Setosa", "Versicolor", "Virginica"), label = "Species",
order = c("Virginica", "Setosa", "Versicolor"))) \%>\%
select(-Species)
nt_describe(iris)

}
