% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{std_barplot}
\alias{std_barplot}
\title{Standard barplot}
\usage{
std_barplot(var, var.label, ylab, ...)
}
\arguments{
\item{var}{a character vector.}

\item{var.label}{a character value specifying the variable label.}

\item{ylab}{a character value specifying the y axis label.}

\item{...}{additional input arguments that may be used when creating your own function.}
}
\value{
a ggplot object.
}
\description{
A function to plot a barplot without groups.
}
\details{
This function defines the standard barplot without groups to be
plotted by the function \code{\link{nt_barplot}}. It can be modified by the
user. See more details in \code{\link[ggplot2]{geom_bar}}.
}
