% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{std_barplot_group}
\alias{std_barplot_group}
\title{Standard barplot by group}
\usage{
std_barplot_group(var, group, var.label, group.label, ylab, ...)
}
\arguments{
\item{var}{a character vector.}

\item{group}{a character vector.}

\item{var.label}{a character value specifying the variable label.}

\item{group.label}{a character value specifying the group label.}

\item{ylab}{a character value specifying the y axis label.}

\item{...}{additional input arguments that may be used when creating your own function.}
}
\value{
a ggplot object.
}
\description{
A function to plot a barplot with groups.
}
\details{
This function defines the standard barplot with groups to be plotted
by the function \code{\link{nt_barplot}}. It can be modified by the user.
See more details in \code{\link[ggplot2]{geom_bar}}.
}
