% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{std_boxplot_group}
\alias{std_boxplot_group}
\title{Standard boxplot by group}
\usage{
std_boxplot_group(var, group, var.label, group.label)
}
\arguments{
\item{var}{a numeric vector.}

\item{group}{a character vector.}

\item{var.label}{a character value specifying the variable label.}

\item{group.label}{a character value specifying the group label.}
}
\value{
a ggplot object.
}
\description{
A function to plot a boxplot with groups.
}
\details{
This function defines the standard boxplot with groups to be plotted
by the function \code{\link{nt_boxplot}}. It can be modified by the user.
See more details in \code{\link[ggplot2]{geom_boxplot}}.
}
