% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{std_dotplot}
\alias{std_dotplot}
\title{Standard dotplot}
\usage{
std_dotplot(var, var.label, binwidth)
}
\arguments{
\item{var}{a numeric vector.}

\item{var.label}{a character value specifying the variable label.}
}
\value{
a ggplot object.
}
\description{
A function to plot a dotplot without groups.
}
\details{
This function defines the standard dotplot without groups to be
plotted by the function \code{\link{nt_dotplot}}. It can be modified by
the user. See more details in \code{\link[ggplot2]{geom_dotplot}}.
}
