% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{std_km}
\alias{std_km}
\title{Standard Kaplan-Meier curve}
\usage{
std_km(time, status, xlab, ylab, risktable.title)
}
\arguments{
\item{time}{a numeric vector.}

\item{status}{a numeric vector of '0' and '1'.}

\item{xlab}{a character value specifying the x axis label.}

\item{ylab}{a character value specifying the y axis label.}
}
\value{
a ggplot object.
}
\description{
A function to plot a Kaplan-Meier curve without groups.
}
\details{
This function defines the standard of Kaplan-Meier curves
without groups to be plotted by the function \code{\link{nt_km}}.
}
