% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{std_km_group}
\alias{std_km_group}
\title{Standard Kaplan-Meier curve by group}
\usage{
std_km_group(
  time,
  status,
  var,
  var.label,
  xlab,
  ylab,
  time.points,
  risktable.title,
  ...
)
}
\arguments{
\item{time}{a numeric vector.}

\item{status}{a numeric vector of '0' and '1'.}

\item{var}{a character vector.}

\item{var.label}{a character value specifying the group label.}

\item{xlab}{a character value specifying the x axis label.}

\item{ylab}{a character value specifying the y axis label.}
}
\value{
a ggplot object.
}
\description{
A function to plot a Kaplan-Meier curve with groups.
}
\details{
This function defines the standard of Kaplan-Meier curves
with groups to be plotted by the function \code{\link{nt_km}}.
It can be modified by the user.
}
