% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileplot.R
\name{std_profileplot}
\alias{std_profileplot}
\title{Standard profile plot}
\usage{
std_profileplot(var, time, var.label, time.label)
}
\arguments{
\item{var}{a numeric vector.}

\item{time}{a numeric vector of factor.}

\item{var.label}{a character value specifying the variable label.}

\item{time.label}{a character value specifying the time label.}
}
\value{
a ggplot object.
}
\description{
A function to plot a profile plot without groups.
}
\details{
This function defines the standard profile plot without groups to be
plotted by the function \code{\link{nt_profileplot}}. It can be modified by the
user. See more details in \code{\link[ggplot2]{geom_errorbar}}.
}
