% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileplot.R
\name{std_profileplot_group}
\alias{std_profileplot_group}
\title{Standard profile plot by group}
\usage{
std_profileplot_group(var, time, group, var.label, time.label, group.label)
}
\arguments{
\item{var}{a numeric vector.}

\item{time}{a numeric vector of factor.}

\item{group}{a character vector.}

\item{var.label}{a character value specifying the variable label.}

\item{time.label}{a character value specifying the time label.}

\item{group.label}{a character value specifying the group label.}
}
\value{
a ggplot object.
}
\description{
A function to plot a profile plot with groups.
}
\details{
This function defines the standard profile plot with groups to be plotted
by the function \code{\link{nt_profileplot}}. It can be modified by the user.
See more details in \code{\link[ggplot2]{geom_errorbar}}.
}
