% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_kruskal_wallis}
\alias{helper_kruskal_wallis}
\title{Helper for Kruskal-Wallis test}
\usage{
helper_kruskal_wallis(x, g)
}
\arguments{
\item{x}{a numeric vector.}

\item{g}{a factor.}
}
\value{
a list with \code{test} indicating the test that was
performed and its respective \code{p.value}.
}
\description{
It performs the Kruskal-Wallis test.
}
\details{
This function can be modified by the user,
but input and output should be kept the same.
}
