% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_mann_whitney}
\alias{helper_mann_whitney}
\title{Helper for Student's t-test}
\usage{
helper_mann_whitney(x, g, paired, alternative, conf.level)
}
\arguments{
\item{x}{a numeric vector.}

\item{g}{a numeric vector.}

\item{paired}{a logical value.}

\item{alternative}{a character value.}

\item{conf.level}{a numeric value.}
}
\value{
a list with the first element \code{test} as the measure name and the
second element as the \code{p.value} for a given variable.
}
\description{
It calculates percentages and frequencies, concatenating them to
present on a table.
}
\details{
This function can be modified by the user,
but input and output should be kept the same.
}
