% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_sf_test}
\alias{helper_sf_test}
\title{Helper for Shapiro Francia test for Normality}
\usage{
helper_sf_test(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
a list indicating the test and its p-value.
}
\description{
It performs the Shapiro-Francia to test the assumption of normality.
}
