% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{nt_compare_mg}
\alias{nt_compare_mg}
\title{Compare more than two groups}
\usage{
nt_compare_mg(
  data,
  group,
  norm.test = nt_norm_test(test = "sf"),
  var.test = nt_var_test(test = "levene"),
  distr.test = list(nt_anova, nt_welch_anova, nt_kruskal_wallis),
  contrast = "Tukey",
  alternative = "two.sided",
  format = TRUE,
  digits.p = 3,
  digits.ci = 2,
  save = FALSE,
  file = "nt_compare_mg",
  mc = FALSE
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{a data frame with the group variable.}

\item{norm.test}{a function with numeric variable and group variable as input, and object with p.value as output.}

\item{var.test}{a function with numeric variable and group variable as input, and object with p.value as output.}

\item{distr.test}{a list of functions for four possible cases: (1) normality and homoscedasticity,
(2) normality and heteroscedasticity, (3) non-normality and homoscedasticity and (4) normality and heteroscedasticity.}

\item{digits.p}{the number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character value indicating the name of output file in csv format to be saved.}

\item{mc}{a logical value indicating if pairwise comparisons should will be performed.}
}
\description{
Performing comparisons among three or more groups.
}
\details{
If \code{test = "automatic"}, the normality assumption will be verified by
a normality test (Anderson-Daling (\link[nortest]{ad.test}),
Shapiro-Francia (\link[nortest]{sf.test}),
'Kolmogorov-Smirnov (\link[nortest]{lillie.test}),
Cramer-vonMises (\link[nortest]{cvm.test}),
and Pearson (\link[nortest]{pearson.test})) and
Levene test (\link[car]{leveneTest}) will evaluate the assumption of
homocedasticity at a significance level of 0.05.
If the data satisfies both assumptions, then ANOVA is chosen;
if only normality is satisfied, then Welch ANOVA; if only homoscedasticity
or neither assumptions, then Kruskal-Wallis.
}
\examples{
library(magrittr)
data(iris)

iris \%>\% nt_compare_mg(group = Species)

}
