% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{nt_compare_tg}
\alias{nt_compare_tg}
\title{Compare two groups}
\usage{
nt_compare_tg(
  data,
  group,
  alternative = "two.sided",
  norm.test = nt_norm_test(test = "sf"),
  var.test = nt_var_test(test = "levene"),
  distr.test = list(nt_student_t, nt_welch_t, nt_mann_whitney, nt_brunner_munzel),
  paired = FALSE,
  conf.level = 0.95,
  format = TRUE,
  digits.ci = 3,
  digits.p = 5,
  save = FALSE,
  file = "nt_compare_tg"
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{a data frame with the group variable.}

\item{alternative}{a character value indicating the alternative hypothesis,
must be one of "two.sided", "greater" or "less".}

\item{norm.test}{a character value specifying the normality test to be performed.
The options are Anderson-Darling (\code{ad}), Shapiro-Francia (\code{"sf"}),
Kolmogorov-Smirnov (\code{ks}),  Cramer-vonMises (\code{cvm}) and
Pearson (\code{p}). The default is Shapiro-Francia (\code{"sf"}). It is only used if
\code{test = "automatic"}.}

\item{paired}{a logical value indicating whether a paired test should be used.}

\item{conf.level}{a character value specifying the confidence level of the confidence interval for
the difference between the two groups.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{digits.ci}{the number of digits to present the confidence intervals.}

\item{digits.p}{the number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character value indicating the name of output file in csv format to be saved.}

\item{test}{a character value indicating the tests to be performed.
The options are "auto", "par", "npar", "t", "wt", "mw", "bm". See more in details.}
}
\description{
Performing comparisons between two groups.
}
\details{
If \code{test = "auto"}, the normality assumption will be verified by
a normality test (Anderson-Daling (\link[nortest]{ad.test}),
Shapiro-Francia (\link[nortest]{sf.test}),
'Kolmogorov-Smirnov (\link[nortest]{lillie.test}),
Cramer-vonMises (\link[nortest]{cvm.test}),
or Pearson (\link[nortest]{pearson.test})) and
Levene test (\link[car]{leveneTest}) will evaluate the assumption of
homocedasticity at a significance level of 0.05.
If the data satisfies both assumptions, then t-test is chosen;
if only normality is satisfied, then Welch t-test is performed; if only homoscedasticity, then
Mann-Whitney is used; if neither assumptions, then Brunner-Munzel t test is applied;
If \code{test = "par"}, then only parametric tests will be performed; If \code{test = "npar"}, then
only non-parametric tests will be performed; If either "t", "wt", "mw" and "bm", then only t-test,
Welch-t test, Mann-Whitney or Brunner-Munzel test will be performed, respectively.
}
\examples{
library(dplyr)
library(forcats)
data(iris)

iris \%>\% filter(Species != "setosa") \%>\%
mutate(Species = fct_drop(Species)) \%>\%
nt_compare_tg(group = Species)

}
