% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{nt_compare_tg}
\alias{nt_compare_tg}
\title{Compare two groups}
\usage{
nt_compare_tg(data, group, alternative = "two.sided", test = "automatic",
  conf.level = 0.95, paired = FALSE, norm.test = "sf", format = TRUE,
  digits.ci = 3, digits.p = 5, save = FALSE, file = "nt_compare_tg")
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{a data frame with the group variable.}

\item{alternative}{a character value indicating the alternative hypothesis,
must be one of "two.sided", "greater" or "less".}

\item{test}{a character value indicating the tests to be performed.
The options are "automatic", "parametric" and "non-parametric". See more in details.}

\item{conf.level}{a character value specifying the confidence level of the confidence interval for
the difference between the two groups.}

\item{paired}{a logical value indicating whether a paired test should be used.}

\item{norm.test}{a character value specifying the normality test to be performed.
The options are Anderson-Darling (\code{ad}), Shapiro-Francia (\code{"sf"}),
Kolmogorov-Smirnov (\code{ks}),  Cramer-vonMises (\code{cvm}) and
Pearson (\code{p}). The default is Shapiro-Francia (\code{"sf"}). It is only used if
\code{test = "automatic"}.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{digits.ci}{the number of digits to present the confidence intervals.}

\item{digits.p}{the number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character value indicating the name of output file in csv format to be saved.}
}
\description{
Performing comparisons between two groups.
}
\details{
If \code{test = "automatic"}, the normality assumption will be verified by
a normality test (Anderson-Daling (\link[nortest]{ad.test}),
Shapiro-Francia (\link[nortest]{sf.test}),
'Kolmogorov-Smirnov (\link[nortest]{lillie.test}),
Cramer-vonMises (\link[nortest]{cvm.test}),
and Pearson (\link[nortest]{pearson.test})) and
Levene test (\link[car]{leveneTest}) will evaluate the assumption of
homocedasticity at a significance level of 0.05.
If the data satisfies both assumptions, then t-test is chosen;
if only normality is satisfied, then Welch t-test; if only homoscedasticity, then
Mann-Whitney; if neither assumptions, then Brunner-Munzel t test.
}
\examples{
library(magrittr)

iris \%>\% nt_compare_tg(group = Species)

}
