% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.r
\name{nt_describe}
\alias{nt_describe}
\title{Descriptive measures}
\usage{
nt_describe(
  data,
  group = NULL,
  measures = list(nt_mean_sd, nt_median_iqr, nt_median_range),
  digits = 2,
  save = FALSE,
  file = "descriptive_analysis"
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{an optional  with the group variable.}

\item{measures}{a list of functions to summarise quantitative variables. See more in details.}

\item{digits}{a numeric value specifying the number of digits to present the results.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\value{
a data frame with summary for all variables by group.
}
\description{
Calculating a descriptive table for quantitative and
qualitative variables in the usual format to present in scientific article.
}
\examples{
data(iris)

iris \%>\% nt_describe(group = Species)

}
