% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.r
\name{nt_describe}
\alias{nt_describe}
\title{Descriptive measures}
\usage{
nt_describe(
  data,
  group = NULL,
  labels = NULL,
  measures_qt = list(helper_mean_sd, helper_median_iqr, helper_median_range,
    helper_missing),
  measures_ql = list(helper_count_perc),
  digits = 2,
  save = FALSE,
  file = "descriptive_analysis",
  ...
)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{an optional  with the group variable.}

\item{labels}{a list of labels with components given by their variable names.}

\item{measures_qt}{a list of functions to summarise quantitative variables. See more in details.}

\item{measures_ql}{a list of functions to summarise qualitative variables. See more in details.}

\item{digits}{a numeric value specifying the number of digits to present the results.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\value{
a data frame with summary for all variables by group.
}
\description{
It calculates a descriptive table for quantitative and
qualitative variables in a publishable format.
}
\details{
For quantitative variables, mean +/- sd, median (quantile 0.25 - quantile 0.75),
median (minimum - maximum) and number of missing observations are calculated using the
functions \code{\link[ntimes]{helper_mean_sd}}, \code{\link[ntimes]{helper_median_iqr}},
\code{\link[ntimes]{helper_median_range}} and \code{\link[ntimes]{helper_missing}}.
For qualitative variables, percentage (frequency) is calculated using
\code{\link[ntimes]{helper_perc_count}}.
}
\examples{
data(iris)

iris |> nt_describe(group = Species,
                    labels = list(Sepal.Length = "Sepal Length",
                                  Sepal.Width = "Sepal Width",
                                  Petal.Length = "Petal Length",
                                  Petal.Width = "Petal Width"))

}
