% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.r
\name{nt_describe}
\alias{nt_describe}
\title{Descriptive measures}
\usage{
nt_describe(data, group = NULL, measures = c("mean.sd", "median.quantiles",
  "median.range"), digits = 2, save = FALSE,
  file = "descriptive_analysis")
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{an optional  with the group variable.}

\item{measures}{a character vector indicating which measures should be presented.}

\item{digits}{a numeric value specifying the number of digits to present the results.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\value{
A data frame with mean, standard deviation, median, quantile 25\%,
quantile 75\%, minimum, maximum, sample size and missing data for
quantitative variables, and frequency, percentage, sample size and
missing data for qualitative variables.
}
\description{
Calculating a descriptive table for quantitative and
qualitative variables in the usual format to present in scientific article.
}
\examples{
library(dplyr)
library(magrittr)

iris <- iris \%>\%
 mutate(Species = ql_var(Species,
                         from = c("setosa", "versicolor", "virginica"),
                         to = c("Setosa", "Versicolor", "Virginica"),
                         order = c("Virginica", "Setosa", "Versicolor"))) \%>\%
iris \%>\% nt_describe(group = Species)

}
