% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{nt_multiple_cox}
\alias{nt_multiple_cox}
\title{Proportional Hazards Cox regression table}
\usage{
nt_multiple_cox(
  fit,
  ci.type = "lr",
  user.contrast = NULL,
  user.contrast.interaction = NULL,
  format = TRUE,
  digits = 2,
  digits.p = 3,
  save = FALSE,
  file = "nt_multiple_cox"
)
}
\arguments{
\item{fit}{a fitted model.}

\item{ci.type}{a character value indicating the procedure to calculate confidence intervals: likelihood ratio (\code{lr}) or wald (\code{wald}).}

\item{user.contrast}{a variable named list of numerical vectors indicating contrast for a covariate.}

\item{user.contrast.interaction}{a variable named list of numerical vectors indicating a contrast for interaction.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{digits}{a numerical value defining of digits to present the results.}

\item{digits.p}{a numerical value defining number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\description{
Tabulating results from fitted Proportional Hazards Cox models.
}
\examples{
library(survival)
library(magrittr)
library(dplyr)

data(ovarian)
ovarian_nt <- ovarian \%>\% mutate(resid.ds = ql_var(resid.ds,
                                               from = 1:2,
                                               to = c("no", "yes"),
                                               label = "Residual Disease"),
                             ecog.ps = ql_var(ecog.ps,
                                              from = 1:2,
                                              to = c("I", "II"),
                                              label = "ECOG-PS"),
                             rx = ql_var(rx,
                                         from = 1:2,
                                         to = c("t1", "t2"),
                                         label = "Treatment"),
                             age = qt_var(age,
                                          label = "Age"))


fit <- coxph(Surv(futime, fustat) ~ age + ecog.ps*rx, data = ovarian_nt)

nt_multiple_cox(fit)

}
