% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{nt_multiple_cox}
\alias{nt_multiple_cox}
\title{Proportional Hazards Cox regression table}
\usage{
nt_multiple_cox(fit.list, fit.labels = NULL, type = "hr",
  format = TRUE, digits = 2, digits.p = 3, save = FALSE,
  file = "nt_multiple_cox")
}
\arguments{
\item{fit.list}{a list of fitted models.}

\item{fit.labels}{a character vector labelling the models in \code{fit.list}}

\item{type}{a character value indicating either hazard ratio (\code{hr}) or coefficients (\code{coef}) will be shown in the output.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{digits}{a numerical value defining of digits to present the results.}

\item{digits.p}{a numerical value defining number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\description{
Tabulating results from fitted Proportional Hazards Cox models.
}
\examples{
library(survival)
library(magrittr)
library(dplyr)

data(ovarian)
ovarian_nt <- ovarian \%>\% mutate(resid.ds = ql_var(resid.ds,
                                               from = 1:2,
                                               to = c("no", "yes"),
                                               label = "Residual Disease"),
                             ecog.ps = ql_var(ecog.ps,
                                              from = 1:2,
                                              to = c("I", "II"),
                                              label = "ECOG-PS"),
                             rx = ql_var(rx,
                                         from = 1:2,
                                         to = c("t1", "t2"),
                                         label = "Treatment"),
                             age = qt_var(age,
                                          label = "Age"))

fit.list <- list()

fit.list[[1]] <- coxph(Surv(futime, fustat) ~ ecog.ps, data = ovarian_nt)
fit.list[[2]] <- coxph(Surv(futime, fustat) ~ resid.ds + rx, data = ovarian_nt)
fit.list[[3]] <- coxph(Surv(futime, fustat) ~ age + ecog.ps*rx, data = ovarian_nt)

nt_multiple_cox(fit.list)

}
