% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{nt_multiple_cox}
\alias{nt_multiple_cox}
\title{Proportional Hazards Cox regression table}
\usage{
nt_multiple_cox(fit.list, format = FALSE, digits = 2, digits.p = 3,
  save = FALSE, file = "nt_multiple_cox")
}
\arguments{
\item{fit.list}{a list of fitted models.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}

\item{data}{a data frame containing the variables used to fit the models listed in fit.list.}
}
\description{
Tabulating results from fitted Proportional Hazards Cox models.
}
\examples{
library(survival)
library(magrittr)
library(dplyr)

data(ovarian)
ovarian_nt <- ovarian \%>\% mutate(resid.ds = ql_var(resid.ds,
                                               from = 1:2,
                                               to = c("no", "yes"),
                                               label = "Residual Disease"),
                             ecog.ps = ql_var(ecog.ps,
                                              from = 1:2,
                                              to = c("I", "II"),
                                              label = "ECOG-PS"),
                             rx = ql_var(rx,
                                         from = 1:2,
                                         to = c("t1", "t2"),
                                         label = "Treatment"),
                             age = qt_var(age,
                                          label = "Age"))

fit.list <- list()

fit.list[[1]] <- coxph(Surv(futime, fustat) ~ ecog.ps, data = ovarian_nt)
fit.list[[2]] <- coxph(Surv(futime, fustat) ~ resid.ds + rx, data = ovarian_nt)
fit.list[[3]] <- coxph(Surv(futime, fustat) ~ age + ecog.ps*rx, data = ovarian_nt)

nt_multiple_cox(fit.list)

}
