% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_regression.R
\name{nt_multiple_logistic}
\alias{nt_multiple_logistic}
\title{Logistic regression table}
\usage{
nt_multiple_logistic(
  fit,
  ci.type = "lr",
  user.contrast = NULL,
  user.contrast.interaction = NULL,
  format = TRUE,
  digits = 2,
  digits.p = 3,
  save = FALSE,
  file = "nt_multiple_logistic"
)
}
\arguments{
\item{fit}{a fitted model.}

\item{ci.type}{a character value indicating the procedure to calculate confidence intervals: likelihood ratio (\code{lr}) or wald (\code{wald}).}

\item{user.contrast}{a variable named list of numerical vectors indicating contrast for a covariate.}

\item{user.contrast.interaction}{a variable named list of numerical vectors indicating a contrast for interaction.}

\item{format}{a logical value indicating whether the output should be formatted.}

\item{digits}{a numerical value defining of digits to present the results.}

\item{digits.p}{a numerical value defining number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\description{
Tabulating results from Logistic models.
}
\examples{
library(titanic)
library(magrittr)
library(dplyr)

data(titanic_train)
dt <- titanic_train \%>\% mutate(Sex = ql_var(Sex,
                                           from = c("male", "female"),
                                           to = c("Male", "Female")),
                              Pclass = ql_var(Pclass,
                                              from = 1:3,
                                              to = c("I", "II", "III"),
                                              label = "Passenger Class"))

fit <- glm(Survived ~ Age + Sex + Pclass, data = dt, family = "binomial")

nt_multiple_logistic(fit)

}
