% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{nt_norm_test}
\alias{nt_norm_test}
\title{Normality tests}
\usage{
nt_norm_test(data, group = NULL, norm.test = helper_sf_test)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{group}{an optional character indicating the group variable.}

\item{norm.test}{a function with a numeric vector as input and a list as output containing an object named \code{p.value} similar to \link[ntimes]{helper_sf_test}.}
}
\value{
\code{tab} a table of p-values for each of the five normality tests.

\code{plot} a Quantile-Quantile plot.
}
\description{
Perform Anderson-Darling, Shapiro-Francia,
Kolmogorov-Smirnov, Cramer-vonMises, and Pearson normality tests
for several variables. In addition, it also plots a Quantile-Quantile plot and a histogram.
}
\examples{
library(magrittr)
data(iris)

iris \%>\% select(-Species) \%>\% nt_norm_test()
iris \%>\% nt_norm_test(group = Species)

}
