% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression.R
\name{nt_simple_cox}
\alias{nt_simple_cox}
\title{Simple Cox Regression}
\usage{
nt_simple_cox(data, time, status, ..., cluster = FALSE, strata = NULL,
  digits = 2, digits.p = 3, save = FALSE, file = "simple_cox",
  format = TRUE)
}
\arguments{
\item{data}{a data frame with the variables.}

\item{time}{a numeric vector with the follow-up time.}

\item{status}{a numeric vector indicating status, 0 = censored, 1 = event at time.}

\item{strata}{a character vector containing the strata.}

\item{digits}{a numerical value defining of digits to present the results.}

\item{digits.p}{a numerical value defining number of digits to present the p-values.}

\item{save}{a logical value indicating whether the output should be saved as a csv file.}

\item{file}{a character indicating the name of output file in csv format to be saved.}
}
\description{
Performing simple Cox regression.
}
\examples{
library(survival)
library(dplyr)
library(magrittr)
data(ovarian)

ovarian_nt <- ovarian \%>\% mutate(age = qt_var(age, label = "Age"),
                                resid.ds = ql_var(resid.ds,
                                                  from = 1:2,
                                                  to = c("no", "yes")),
                                ecog.ps = ql_var(ecog.ps,
                                                 from = 1:2,
                                                 to = c("I", "II")),
                                rx = ql_var(rx,
                                            from = 1:2,
                                            to = c("t1", "t2")))
ovarian_nt \%>\% nt_simple_cox(time = futime, status = fustat)

}
