% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{qt_var}
\alias{qt_var}
\title{Format qualitative variables}
\usage{
qt_var(var, label = NULL, unit = NULL)
}
\arguments{
\item{var}{vector to be formatted.}

\item{label}{an optional character to be attributed as a label to \code{var}.}

\item{unit}{an optional character to be attributed as a unit to \code{var}.}
}
\value{
a vector with an attribute label if \code{label} is provided,
and an attribute unit if \code{unit} is provided.
}
\description{
Recode factors, change the order of levels and
add a label to the qualitative variable.
}
\examples{
library(dplyr)
library(magrittr)
data(iris)

iris_nt <- iris \%>\%
mutate(Sepal.Length = qt_var(Sepal.Length, label = "Sepal Length", unit = "cm"),
      Sepal.Width = qt_var(Sepal.Width, label = "Sepal Width", unit = "cm"),
      Petal.Length = qt_var(Petal.Length, label = "Petal Length", unit = "cm"),
      Petal.Width = qt_var(Petal.Width, label = "Petal Length", unit = "cm"))
iris_nt \%>\% nt_describe(group = Species)

}
