% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{std_dotplot_group}
\alias{std_dotplot_group}
\title{Standard dotplot by group}
\usage{
std_dotplot_group(var, group, var.label, group.label, binwidth)
}
\arguments{
\item{var}{a numeric vector.}

\item{group}{a character vector.}

\item{var.label}{a character value specifying the variable label.}

\item{group.label}{a character value specifying the group label.}
}
\value{
a ggplot object.
}
\description{
A function to plot a dotplot with groups.
}
\details{
This function defines the standard dotplot with groups to be
plotted by the function \code{\link{nt_dotplot}}. It can be modified by
the user. See more details in \code{\link[ggplot2]{geom_dotplot}}.
}
